+++ packages/bitcoin-module/CHANGELOG.md
@@ -1,5 +1,31 @@
 # @avalabs/bitcoin-module
 
+## 0.11.9
+
+### Patch Changes
+
+- @avalabs/vm-module-types@0.11.9
+
+## 0.11.8
+
+### Patch Changes
+
+- @avalabs/vm-module-types@0.11.8
+
+## 0.11.7
+
+### Patch Changes
+
+- ed2e123: ceil gas limit
+  - @avalabs/vm-module-types@0.11.7
+
+## 0.11.6
+
+### Patch Changes
+
+- ed2e123: ceil gas limit
+  - @avalabs/vm-module-types@0.11.6
+
 ## 0.11.5
 
 ### Patch Changes